requirements.txt
boto3==1.42.3
botocore==1.42.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
faiss-cpu==1.13.0
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
fuzzywuzzy==0.18.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
Levenshtein==0.27.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
